Make sdimacs files parsable from ABC command line
ABC supports the following syntax in its command-line interface:
`bin/abc -c '<commands>' <input>`
(In the past we used `bin/abc -c 'read <input>; <commands>'`.)

Such syntax makes the usage of BenchExec easier,
because the commands (rundefinition) and the input (task) are separate.

This commit modifies the parsing of ABC such that when it encounters a sdimacs file,
it appends the file name to the user commands.
This makes the above syntax also possible for SSAT solving.