Bugfix for issue #12
Bugfix: multiplier for unit clauses

If the matrix has unit clauses of random variables,
their probabilities are not included in the weighted model counting,
because the SAT solver does not store a unit clause in its clause set.

Therefore, we have to adjust the computed satisfying probability with a multiplier,
which is product of the probabilities of all unit random literals.

Bugfix: stored _unitClause might be incomplete

Counterexample:
Suppose clause "-1 2 0" is read before "1 0",
_unitClause will only store "1 0" but ignore "2 0".

Instead, this commit uses solver's data member assigns to decide unit clauses in the matrix.