Bugfix for issue #11
Small output formatting

Bugfix: consider unit clauses in pure-literal assertion

The original code for pure literal assertion did not take unit clauses into account.
If an existential variable x at lv=0 is pure, it is asserted in the selection solver.

The pure-literal assertion code reads the clauses and finds that only !x appears in the matrix.
Therefore, it asserts !x in the selection solver, leading to an UNSAT and zero satisfying probability.

Solution: do not assert any variable which is a unit clause.