Development for issue #10 (early termination)
The solver will start to compute bounds upon `SIGTERM` and will be killed upon `SIGKILL`.
These signals will be sent by BenchExec.
(It sends `SIGTERM` at timelimit and `SIGKILL` at hardtimelimit.)

To avoid BDD from crashing when computing the bounds,
a limit of 60000 cubes is imposed.
If BDD is used and there are more than 60000 cubes, the bound will not be computed.